requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==1.12.0
httpx==0.27.2
instructor==0.4.5